pyproject.toml
[project]
name = "comcer"
version = "0.1.0"
description = ""

authors = [
    {name = "JeanPaulSB",email = "jeanpaulsierraboom@gmail.com"}
]
readme = "README.md"
package-mode = false
requires-python = ">=3.9, <=3.14"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "google-api-python-client (>=2.174.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pypdf (>=5.7.0,<6.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pyexcel (>=0.7.3,<0.8.0)",
    "pyexcel-xls (>=0.7.1,<0.8.0)",
    "pyexcel-xlsx (>=0.6.1,<0.7.0)",
    "reportlab (>=4.4.2,<5.0.0)",
    "oauth2client (>=4.1.3,<5.0.0)",
    "gspread-dataframe (>=4.0.0,<5.0.0)"
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
